Refactor PokeApi Error to need any external parameters
This is a step towards simplifying the error story in the PokeApi:
Move towards the same style like the Translation API and have a
single From<...> implementation to convert form a reqwest Error.
Should this API get more complex in the future, then I can still
fall back on an explicit `.map_err(...)` implementation.
At the moment, there is no new context gained and some `log::error!`
macros sprinkled in.