Compromise legibility vs valuable information and log at the server level
This is truly a compromise: if we were to extract more valuable
information at the lower levels (e.g. pokeapi or translation api) I
would definitly want to log it there. This could be timings, or events,
or even details from responses.
A further aspect is complexity: There are very limited code paths in
those lower modules -they are a pipeline, effectively.
The only place we have complex branching logic is at the server level,
is it can establish the context better.

This would change if I used a more advanced library like Tokios `trace`
instead of plain ol' `log`.